+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-06-11
+
+### Changed
+
+- AWS QuickSight embedded dashboards to create individual users upon request.
+
 ## 2020-06-10
 
 ### Removed